Move fullscreen to separate file
